+++ package.json
@@ -19,7 +19,7 @@
     "ckeditor5-ui": "ckeditor/ckeditor5-ui",
     "ckeditor5-ui-default": "ckeditor/ckeditor5-ui-default",
     "ckeditor5-undo": "ckeditor/ckeditor5-undo",
-    "ckeditor5-utils": "ckeditor/ckeditor5-utils",
+    "ckeditor5-utils": "ckeditor/ckeditor5-utils#t/39",
     "requirejs": "Reinmar/requirejs"
   },
   "devDependencies": {
