+++ src/editor/editor.js
@@ -1,0 +1,166 @@
+/**
+ * @license Copyright (c) 2003-2016, CKSource - Frederico Knabben. All rights reserved.
+ * For licensing, see LICENSE.md.
+ */
+
+'use strict';
+
+import EmitterMixin from '../utils/emittermixin.js';
+import Config from '../utils/config.js';
+import PluginCollection from '../plugincollection.js';
+import Locale from '../utils/locale.js';
+import DataController from '../engine/datacontroller.js';
+import Document from '../engine/model/document.js';
+
+import CKEditorError from '../utils/ckeditorerror.js';
+import isArray from '../utils/lib/lodash/isArray.js';
+import mix from '../utils/mix.js';
+
+/**
+ * Represents a single editor instance.
+ *
+ * @memberOf ckeditor5
+ * @mixes utils.EmitterMixin
+ */
+export default class Editor {
+	/**
+	 * Creates a new instance of the Editor class.
+	 *
+	 * @param {Object} config The editor config.
+	 */
+	constructor( config ) {
+		/**
+		 * Holds all configurations specific to this editor instance.
+		 *
+		 * @readonly
+		 * @member {utils.Config} ckeditor5.Editor#config
+		 */
+		this.config = new Config( config );
+
+		/**
+		 * The plugins loaded and in use by this editor instance.
+		 *
+		 * @readonly
+		 * @member {ckeditor5.PluginCollection} ckeditor5.Editor#plugins
+		 */
+		this.plugins = new PluginCollection( this );
+
+		/**
+		 * Commands registered to the editor.
+		 *
+		 * @readonly
+		 * @member {Map.<ckeditor5.command.Command>} ckeditor5.Editor#commands
+		 */
+		this.commands = new Map();
+
+		/**
+		 * @readonly
+		 * @member {utils.Locale} ckeditor5.Editor#locale
+		 */
+		this.locale = new Locale( this.config.get( 'lang' ) );
+
+		/**
+		 * Shorthand for {@link utils.Locale#t}.
+		 *
+		 * @see utils.Locale#t
+		 * @method ckeditor5.Editor#t
+		 */
+		this.t = this.locale.t;
+
+		/**
+		 * Tree Model document managed by this editor.
+		 *
+		 * @readonly
+		 * @member {engine.model.Document} ckeditor5.Editor#document
+		 */
+		this.document = new Document();
+
+		/**
+		 * Instance of the {@link engine.DataController data controller}.
+		 *
+		 * @readonly
+		 * @member {engine.DataController} ckeditor5.Editor#data
+		 */
+		this.data = new DataController( this.document );
+	}
+
+	/**
+	 * TODO
+	 *
+	 * @returns {Promise} A promise which resolves once the initialization is completed.
+	 */
+	initPlugins() {
+		const that = this;
+		const config = this.config;
+
+		return loadPlugins()
+			.then( initPlugins );
+
+		function loadPlugins() {
+			let plugins = config.features || [];
+
+			// Handle features passed as a string.
+			if ( !isArray( plugins ) ) {
+				plugins = plugins.split( ',' );
+			}
+
+			return that.plugins.load( plugins );
+		}
+
+		function initPlugins( loadedPlugins ) {
+			return loadedPlugins.reduce( ( promise, plugin ) => {
+				return promise.then( plugin.init.bind( plugin ) );
+			}, Promise.resolve() );
+		}
+	}
+
+	/**
+	 * Destroys the editor instance, releasing all resources used by it. If the editor replaced an element, the
+	 * element will be recovered.
+	 *
+	 * @fires ckeditor5.Editor#destroy
+	 * @returns {Promise} A promise that resolves once the editor instance is fully destroyed.
+	 */
+	destroy() {
+		this.fire( 'destroy' );
+		this.stopListening();
+
+		// Note: The destruction order should be the reverse of the initialization order.
+		return Promise.resolve()
+			.then( () => {
+				this.document.destroy();
+				this.editing.destroy();
+				this.data.destroy();
+			} );
+	}
+
+	/**
+	 * Executes specified command with given parameter.
+	 *
+	 * @param {String} commandName Name of command to execute.
+	 * @param {*} [commandParam] If set, command will be executed with this parameter.
+	 */
+	execute( commandName, commandParam ) {
+		let command = this.commands.get( commandName );
+
+		if ( !command ) {
+			/**
+			 * Specified command has not been added to the editor.
+			 *
+			 * @error editor-command-not-found
+			 */
+			throw new CKEditorError( 'editor-command-not-found: Specified command has not been added to the editor.' );
+		}
+
+		command._execute( commandParam );
+	}
+}
+
+mix( Editor, EmitterMixin );
+
+/**
+ * Fired when this editor instance is destroyed. The editor at this point is not usable and this event should be used to
+ * perform the clean-up in any plugin.
+ *
+ * @event ckeditor5.Editor#destroy
+ */
